// src/usuario/dto/usuario.adicionar.dto.ts
export interface UsuarioAdicionarDto{
    id: number;
    nome: string;
    login: string;
    senha: string;
    email: string;
    telefone: string;
    ocupacao: string;
    perfil: string;
    codigo: string;
    disc: string;
    saude: string;
    descricaoAnalise: string;
    descricaoSaude: string;
    papel: string;
};